% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.garch.tse.R
\docType{package}
\name{display}
\alias{tswge2}
\alias{display}
\title{GARCH Model Display Functions}
\description{
Functions for displaying GARCH model comparison results
and coefficient tables in various formats.
}
\author{
\strong{Maintainer}: Bivin Sadler \email{bsadler@smu.edu}

Authors:
\itemize{
  \item Wayne Woodward
}

Other contributors:
\itemize{
  \item Kyle Evans [contributor]
}

}
\keyword{internal}
