% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.tse.R
\docType{package}
\name{generators}
\alias{generators}
\title{Innovation and Process Generators}
\description{
Factory functions for creating innovation generators and
stochastic process generators for time series simulation.
}
\author{
\strong{Maintainer}: Bivin Sadler \email{bsadler@smu.edu}

Authors:
\itemize{
  \item Wayne Woodward
}

Other contributors:
\itemize{
  \item Kyle Evans [contributor]
}

}
\keyword{internal}
