% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.generators.tse.R
\name{make.stat.co.tse}
\alias{make.stat.co.tse}
\title{Create Cochrane-Orcutt Statistic Function}
\usage{
make.stat.co.tse(maxp = 5)
}
\arguments{
\item{maxp}{Integer. Maximum AR order to consider for the error model.
Default is 5.}
}
\value{
A function that takes a numeric vector \code{x} and returns
the Cochrane-Orcutt t-statistic for the trend coefficient.
}
\description{
Creates a function that computes the Cochrane-Orcutt t-statistic for
testing trend in a time series with autocorrelated errors.
}
\examples{
\dontrun{
stat_fn <- make.stat.co.tse(maxp = 3)
result <- wbg.boot.test.tse(x, stat_fn = stat_fn)
}

}
\seealso{
\code{\link{wbg.boot.test.tse}}, \code{\link{co.wge}}
}
