% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.aruma.tse.R
\name{plot.aruma}
\alias{plot.aruma}
\title{Plot Method for aruma Objects}
\usage{
\method{plot}{aruma}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"aruma"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Plot Method for aruma Objects
}
