% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.aruma.tse.R
\name{print.aruma}
\alias{print.aruma}
\title{Print method for aruma objects}
\usage{
\method{print}{aruma}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"aruma"} from \code{\link{gen.aruma.tse}}.}

\item{...}{Additional arguments passed to print methods (currently ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for aruma objects
}
\seealso{
\code{\link{gen.aruma.tse}}
}
