% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.garch.tse.R
\name{print.garch.comparison.tse}
\alias{print.garch.comparison.tse}
\title{Print method for garch.comparison.tse}
\usage{
\method{print}{garch.comparison.tse}(x, ...)
}
\arguments{
\item{x}{A garch.comparison.tse object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for garch.comparison.tse
}
