% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.garch.tse.R
\name{summary.garch.comparison.tse}
\alias{summary.garch.comparison.tse}
\title{Summary method for garch.comparison.tse}
\usage{
\method{summary}{garch.comparison.tse}(object, ...)
}
\arguments{
\item{object}{A garch.comparison.tse object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns object
}
\description{
Summary method for garch.comparison.tse
}
