% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tswge-package.R
\docType{package}
\name{tswge2-package}
\alias{tswge2-package}
\title{tswge2: Time Series for Data Science}
\description{
Accompanies the texts Time Series for Data Science with R by Woodward, Sadler and Robertson & Applied Time Series Analysis with R, 2nd edition by Woodward, Gray, and Elliott. It is helpful for data analysis and for time series instruction. Includes tools for GARCH model comparison and diagnostics.
}
\author{
\strong{Maintainer}: Bivin Sadler \email{bsadler@smu.edu}

Authors:
\itemize{
  \item Wayne Woodward
}

Other contributors:
\itemize{
  \item Kyle Evans [contributor]
}

}
\keyword{internal}
