% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.aruma.tse.R
\name{plot.aruma}
\alias{plot.aruma}
\title{Plot method for aruma objects}
\usage{
\method{plot}{aruma}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"aruma"} from \code{\link{gen.aruma.tse}}.}

\item{...}{Additional arguments passed to print methods (currently ignored).}
}
\value{
Invisibly returns the \code{ggplot} object.
}
\description{
Plot method for aruma objects
}
\seealso{
\code{\link{gen.aruma.tse}}
}
