% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.aruma.tse.R
\name{print.aruma}
\alias{print.aruma}
\title{Print Method for aruma Objects}
\usage{
\method{print}{aruma}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"aruma"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for aruma Objects
}
